Moving functions reading "input" structure from the outfile from joutstructures.py to jdftxoutfileslice.py, making the "init_struc" an initialization argument for a JOutStructures instead of something constructed during initialization, reducing the number of instances where object indexes are referenced and increasing the number of attributes explicitly set as None to avoid later references to unset attributes
